Fixed Algorithm Stats Output

+++ SortingAlgorithm.c
@@ -70,8 +70,8 @@ int main() {
 	*/
 
 	int swaps = 0;
-	int comparisons = 0;
-	int accesses = 0;
+	long comparisons = 0;
+	long accesses = 0;
 
 	clock_t start = clock(); // Get Start Timestamp
 
@@ -114,7 +114,7 @@ int main() {
 	// ------- Sorting Algorithm Stats ------- //
 
 	printf(
-		"  Comparisons: %i  |  Swaps: %i  |  Array Accesses: %i  |  Algorithm Time: %.4fs\n\n", 
+		"  Comparisons: %ld  |  Swaps: %i  |  Array Accesses: %ld  |  Algorithm Time: %.4fs\n\n", 
 		comparisons, swaps, accesses, (float)(duration / 1000)
 	);
